requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.3
entrypoints==0.3
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.10.0
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
